restructured project into a distributable Python package
