+++ manolo_bot/__init__.py
@@ -1,0 +1,5 @@
+from manolo_bot.ai.llmagent import LLMAgent
+from manolo_bot.ai.llmbot import LLMBot, LLMBuilder
+from manolo_bot.storage.base import BaseMessagesStorage
+
+__all__ = ["LLMBot", "LLMBuilder", "LLMAgent", "BaseMessagesStorage"]
