+++ manolo_bot/ai/config.py
