+++ manolo_bot/storage/base.py
