--- pytest.ini
@@ -1,8 +1,0 @@
-[pytest]
-asyncio_mode = auto
-asyncio_default_fixture_loop_scope = function
-testpaths = tests
-python_files = test_*.py
-python_classes = Test*
-python_functions = test_*
-
