+++ requirements-dev.txt
@@ -2,34 +2,19 @@
 #    uv export --format requirements-txt --only-dev --no-hashes
 cfgv==3.5.0
     # via pre-commit
-colorama==0.4.6 ; sys_platform == 'win32'
-    # via pytest
 distlib==0.4.0
     # via virtualenv
 filelock==3.20.3
     # via virtualenv
 identify==2.6.16
     # via pre-commit
-iniconfig==2.3.0
-    # via pytest
 nodeenv==1.10.0
     # via pre-commit
-packaging==25.0
-    # via pytest
 platformdirs==4.5.1
     # via virtualenv
-pluggy==1.6.0
-    # via pytest
 pre-commit==4.5.1
-pygments==2.19.2
-    # via pytest
-pytest==9.0.2
-    # via pytest-asyncio
-pytest-asyncio==1.3.0
 pyyaml==6.0.3
     # via pre-commit
 ruff==0.14.14
-typing-extensions==4.15.0 ; python_full_version < '3.13'
-    # via pytest-asyncio
 virtualenv==20.36.1
     # via pre-commit
